Modifies the compile line, final commit
There is some cleanup to do in the code. The measureAndExecute method
is not needed but since it took me a while, and some external help, to
get it working I might as well leave it.

Removing some of this unnecessary code would avoid the need for some
of the compiler flags.